# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "druthers.io API local",
    "description": "API for druthers.io \u2014 track and rank the movies, TV, books, and games you actually care about.",
    "contact": {
      "name": "Druthers",
      "url": "https://www.druthers.io/"
    },
    "license": {
      "name": "GPL-3.0",
      "url": "https://www.gnu.org/licenses/gpl-3.0.html"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "https://api.druthers.io",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/v1/users/me/friends/requests": {
      "get": {
        "tags": [
          "Friends"
        ],
        "summary": "List Friend Requests",
        "operationId": "list_friend_requests_v1_users_me_friends_requests_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutPendingFriendRequests"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Friends"
        ],
        "summary": "Send Friend Request",
        "description": "Send a friend request to an exact handle.\n\nAlways 202 unless the caller is already a party to the relationship (or\nis the target) \u2014 see the module docstring for why the shape of the\nnon-answers matters as much as the answer.",
        "operationId": "send_friend_request_v1_users_me_friends_requests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InFriendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutFriendAck"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/v1/users/me/friends/requests/{request_id}/accept": {
      "put": {
        "tags": [
          "Friends"
        ],
        "summary": "Accept Friend Request",
        "description": "Accept an incoming request; from here both sides count as friends.",
        "operationId": "accept_friend_request_v1_users_me_friends_requests__request_id__accept_put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutFriend"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/users/me/friends/requests/{request_id}/decline": {
      "put": {
        "tags": [
          "Friends"
        ],
        "summary": "Decline Friend Request",
        "description": "Decline an incoming request.\n\nDeletes the row instead of recording a refusal: nobody should be able to\nread a standing \"no\" off the graph, and the pair can try again later.",
        "operationId": "decline_friend_request_v1_users_me_friends_requests__request_id__decline_put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutFriendAck"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/users/me/friends/requests/{request_id}": {
      "delete": {
        "tags": [
          "Friends"
        ],
        "summary": "Cancel Friend Request",
        "description": "Withdraw a request the caller sent, before it is answered.",
        "operationId": "cancel_friend_request_v1_users_me_friends_requests__request_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/users/me/friends": {
      "get": {
        "tags": [
          "Friends"
        ],
        "summary": "List Friends",
        "operationId": "list_friends_v1_users_me_friends_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OutFriend"
                  },
                  "type": "array",
                  "title": "Response List Friends V1 Users Me Friends Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/v1/users/me/friends/{friendship_id}": {
      "delete": {
        "tags": [
          "Friends"
        ],
        "summary": "Unfriend",
        "description": "End a friendship. Symmetric: either side may do this, and one row going\naway ends it for both \u2014 there is no second row to leave behind.",
        "operationId": "unfriend_v1_users_me_friends__friendship_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "friendship_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Friendship Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InFriendRequest": {
        "properties": {
          "handle": {
            "type": "string",
            "maxLength": 30,
            "minLength": 1,
            "title": "Handle"
          }
        },
        "type": "object",
        "required": [
          "handle"
        ],
        "title": "InFriendRequest",
        "description": "Request body for sending a friend request (#275).\n\nA handle and nothing else: there is no endpoint that lists or searches\nusers, so an exact handle someone told you is the only way to reach them."
      },
...
      "OutFriend": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user": {
            "$ref": "#/components/schemas/OutFriendUser"
          },
          "friends_since": {
            "type": "string",
            "format": "date-time",
            "title": "Friends Since"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user",
          "friends_since"
        ],
        "title": "OutFriend",
        "description": "One accepted friendship. ``id`` addresses the friendship, not the user."
      },
      "OutFriendAck": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "OutFriendAck",
        "description": "Deliberately contentless acknowledgement of a friend-graph write.\n\nSend and decline both answer with this, and send answers with *exactly*\nthis whether or not the handle resolved to anybody \u2014 see\n``router_friends`` for why the response must not vary."
      },
      "OutFriendRequest": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user": {
            "$ref": "#/components/schemas/OutFriendUser"
          },
          "requested_at": {
            "type": "string",
            "format": "date-time",
            "title": "Requested At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user",
          "requested_at"
        ],
        "title": "OutFriendRequest",
        "description": "One pending request, from the perspective of whoever is listing it."
      },
      "OutFriendUser": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "handle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Handle"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "OutFriendUser",
        "description": "The other party in a friendship, as much of them as a friend may see.\n\nNo email, no visibility settings \u2014 a friendship is not an introduction to\nsomeone's account."
      },
      "OutPendingFriendRequests": {
        "properties": {
          "incoming": {
            "items": {
              "$ref": "#/components/schemas/OutFriendRequest"
            },
            "type": "array",
            "title": "Incoming",
            "default": []
          },
          "outgoing": {
            "items": {
              "$ref": "#/components/schemas/OutFriendRequest"
            },
            "type": "array",
            "title": "Outgoing",
            "default": []
          }
        },
        "type": "object",
        "title": "OutPendingFriendRequests",
        "description": "Both directions in one response.\n\nThe inbox badge and the \"pending\" state of an outgoing request are drawn\ntogether, so splitting these across two endpoints would only ever mean\ntwo calls."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "v1/auth/token"
          }
        }
      }
    }
  },
...
}